% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optdigits.R
\docType{data}
\name{optdigits}
\alias{optdigits}
\title{optdigits}
\format{Data frame with 65 columns}
\source{
Optical Recognition of Handwritten Digits Data
 \url{https://archive.ics.uci.edu/ml/datasets/Optical+Recognition+of+Handwritten+Digits}
}
\usage{
optdigits
}
\description{
Number of Attributes
  64 input+1 class attribute
}
\details{
For Each Attribute:
  All input attributes are integers in the range 0..16.
  The last attribute is the class code 0..9
}
\examples{
data(optdigits)
}
\keyword{datasets}

